Update Beautiful Year.cpp

+++ 800/Beautiful Year.cpp
@@ -3,8 +3,8 @@
 #define ll long long int
 using namespace std;
 bool check(int n){
-    int c[10];
-    for(int i=0; i<10; ++i) c[i]=0;
+    ll c[10];
+    for(ll i=0; i<10; ++i) c[i]=0;
     while(n>0){
         if(c[n%10]++)return false;
         n/=10;
@@ -13,7 +13,7 @@ bool check(int n){
     }
     
     int main(){
-    int x;
+    ll x;
     scanf("%d",&x);
     while(++x) if(check(x)) return 0&printf("%d\n",x);
     return 0;
